requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database (prepararemos para múltiples BD)
sqlalchemy==2.0.35
alembic==1.13.3
pymysql==1.1.1
cryptography==43.0.1

# SQLite support (ya viene con Python, pero por si acaso)
# No requiere instalación adicional

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

# Cloudinary
cloudinary==1.41.0

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Development tools
black==24.8.0
flake8==7.1.1
mypy==1.11.2